Fix wrong index format

+++ web/xatlas-web.cpp
@@ -18,7 +18,7 @@ MeshBufferInfo createMesh(uint32_t vertexCount, uint32_t indexCount, bool normal
   meshDecl->indexCount = indexCount;
 
   meshDecl->indexData = new int[indexCount];
-  meshDecl->indexFormat = xatlas::IndexFormat::UInt16;
+  meshDecl->indexFormat = xatlas::IndexFormat::UInt32;
   meshBufferInfo.indexOffset = (uint32_t)meshDecl->indexData;
 
   meshDecl->vertexPositionData = new float[vertexCount * 3];
@@ -47,7 +47,7 @@ UvMeshBufferInfo createUvMesh(int vertexCount, int indexCount) {
   uvMeshDecl = new xatlas::UvMeshDecl;
   uvMeshDecl->vertexCount = vertexCount;
   uvMeshDecl->indexCount = indexCount;
-  
+
   uvMeshDecl->indexData = new uint16_t[indexCount];
   uvMeshDecl->indexFormat = xatlas::IndexFormat::UInt16;
   uvMeshBufferInfo.indexOffset = (uint32_t)uvMeshDecl->indexData;
